requirements_cron_job_viewer.txt
# Cron Job Viewer Dependencies

# Core Django dependencies (already included in main project)
# django>=4.2.0
# djangorestframework>=3.14.0

# Celery and Redis for task scheduling
celery>=5.3.0
redis>=4.5.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0

# Cron expression parsing
croniter>=1.4.0

# Database drivers (choose based on your database)
# PostgreSQL
psycopg2-binary>=2.9.0

# MySQL (alternative)
# mysqlclient>=2.1.0

# SQLite (included with Python)

# Additional utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Development and testing
# pytest>=7.0.0
# pytest-django>=4.5.0
# factory-boy>=3.2.0
